re organized lambdas
