+++ services/auth/checkAuth.js
