+++ services/hack/hackDetail.js
